+++ lib/exception/negative_number_exception.dart
@@ -1,0 +1,10 @@
+class NegativeNumberException implements Exception {
+  final List<int> negativeNumbers;
+
+  NegativeNumberException(this.negativeNumbers);
+
+  @override
+  String toString() {
+    return 'Negative numbers not allowed: ${negativeNumbers.join(',')}';
+  }
+}