+++ lib/simple_tdd.dart
@@ -1,3 +1,5 @@
+import 'package:simple_tdd/exception/negative_number_exception.dart';
+
 class StringCalculator {
   int add(String? input) {
     if (input == null || input.isEmpty) return 0;
@@ -7,7 +9,14 @@ class StringCalculator {
     final parts = numbers.split(RegExp(delimiterPattern));
 
     // error parsing ? then setting 0
-    final numbersArray = parts.map((e) => int.tryParse(e) ?? 0);
+    final numbersArray = parts.map((e) => int.tryParse(e) ?? 0).where((n) => n <= 1000);
+
+    final negativeNumbers = numbersArray.where((n) => n < 0).toList(); // Numbers bigger than 1000 should be ignored, so adding 2 + 1001 = 2
+
+    if (negativeNumbers.isNotEmpty) {
+      throw NegativeNumberException(negativeNumbers);
+    }
+
     var sum = 0;
     numbersArray.forEach((n) => sum += n);
     return sum;